Add --dryrun flag so that we can check what the script would do.

+++ archivevideos.py
@@ -1,7 +1,8 @@
 import logging
 import urllib
 import os
-import math
+import argparse
+import errno
 import boto3
 import botocore
 from KalturaClient import *
@@ -9,6 +10,9 @@ from KalturaClient.Plugins.Core import *
 from datetime import datetime
 import calendar
 from dateutil.relativedelta import relativedelta
+
+# Whether or not to modify any data
+dryrun = False
 
 # Flavors should be deleted after the video has not been played for this many years
 years2deleteflavors = 0
@@ -24,6 +28,7 @@ downloaddir = "/tmp"
 # Name of S3 Glacier bucket
 s3bucketname = "kalturavids"
 
+# File to be uploaded when all flavors are deleted
 placeholder_file_path = "./placeholder_video.mp4"
 
 # Kaltura KMC connection information, pulled from environment variables
@@ -40,10 +45,25 @@ privileges = "disableentitlement"
 # Logging configuration
 logging.basicConfig(filename='./archivevideos.log',level=logging.INFO, format='%(asctime)s [%(levelname)s] %(message)s')
 
-def getEntriesToArchive():
-       entrylist = ""
-       return entrylist
+def checkConfig():
 
+    # Check for access to S3 bucket
+    try:
+        s3resource = boto3.resource('s3')
+        s3resource.meta.client.head_bucket(Bucket=s3bucketname)
+    except Exception as e:
+        #logging.fatal("Exception type is: %s" % (e.__class__.__name__))
+        #logging.fatal("Exception message: %s" % (e.message))
+        logging.fatal("Cannot access S3 Bucket: %s" % (s3bucketname))
+        logging.fatal("Exception message: %s" % (e.message))
+        logging.fatal("Exiting immediately")
+        exit(errno.ENOENT)
+
+    # Check for existence of placeholder video
+    if not os.path.isfile(placeholder_file_path):
+        logging.fatal("Placeholder video does not exist: %s" % (placeholder_file_path))
+        logging.fatal("Exiting immediately")
+        exit(errno.ENOENT)
 
 def _getSearchFilter(yearssinceplay, tag, categoryid, entryid):
     # Get list
@@ -53,7 +73,7 @@ def _getSearchFilter(yearssinceplay, tag, categoryid, entryid):
     filter.orderBy = "+createdAt"  # Oldest first
     filter.mediaTypeEqual = KalturaMediaType.VIDEO
 
-# Test file
+# Test entry
     filter.idEqual = "1_6cwwzio0"
 
     if entryid is not None:
@@ -125,7 +145,8 @@ def deleteFlavors():
         # But if there is a source video, then delete all other flavors
         else:
           # Delete the flavors
-          _deleteEntryFlavors(entry)
+          _deleteEntryFlavors(entry.id, False)
+
           # Tag the video so that we know that this script deleted the flavors
           _addTag(entry, flavorsdeletedtag)
 
@@ -135,14 +156,22 @@ def deleteFlavors():
       pager.pageIndex += 1
 
 
-def _deleteEntryFlavors(entry):
-          flavorassetswparamslist = client.flavorAsset.getFlavorAssetsWithParams(entry.id)
+def _deleteEntryFlavors(entryid, includesource=False):
 
-          for flavorassetwparams in flavorassetswparamslist:
-            flavorasset = flavorassetwparams.getFlavorAsset()
-            if (flavorasset != None and not flavorasset.getIsOriginal()):
-              logging.info("Deleting flavor: %s from entry: %s" % (flavorasset.id, entry.id))
-              client.flavorAsset.delete(flavorasset.id)
+    flavorassetswparamslist = client.flavorAsset.getFlavorAssetsWithParams(entryid)
+
+    for flavorassetwparams in flavorassetswparamslist:
+        flavorasset = flavorassetwparams.getFlavorAsset()
+
+        if (flavorasset is not None and flavorasset.getIsOriginal() and includesource):
+            logging.info("Deleting source flavor: %s from entry: %s" % (flavorasset.id, entryid))
+            if not dryrun:
+                client.flavorAsset.delete(flavorasset.id)
+
+        elif (flavorasset is not None and not flavorasset.getIsOriginal()):
+            logging.info("Deleting derived flavor: %s from entry: %s" % (flavorasset.id, entryid))
+            if not dryrun:
+                client.flavorAsset.delete(flavorasset.id)
 
 
 def _getSourceFlavor(entry):
@@ -152,7 +181,7 @@ def _getSourceFlavor(entry):
             #print(type(flavorassetwparams).__name__)
             flavorasset = flavorassetwparams.getFlavorAsset()
 
-            if ( flavorasset != None and flavorasset.getIsOriginal()):
+            if ( flavorasset is not None and flavorasset.getIsOriginal()):
               return flavorasset
 
           # If the original wasn't found
@@ -161,7 +190,9 @@ def _getSourceFlavor(entry):
 def _addTag(entry, newtag):
         mediaEntry = KalturaMediaEntry()
         mediaEntry.tags = entry.tags + ", " + newtag
-        client.media.update(entry.id, mediaEntry)
+
+        if not dryrun:
+            client.media.update(entry.id, mediaEntry)
 
 
 def archiveFlavors():
@@ -202,27 +233,28 @@ def archiveFlavors():
             else:
               # Look ahead to see if this entry_id is already in S3, if it does then skip
               if _S3ObjectExists(s3resource, s3bucketname, entry.id):
-                logging.warning("Entry %s already exists in S3!!!" % (entry.id))
+                logging.warning("Source file for entry %s already exists in S3!!!" % (entry.id))
                 nid += 1
                 continue
 
-              videofile = _downloadVideoFile(sourceflavor)
+              logging.info("Archiving entry: %s" % (entry.id))
+              if not dryrun:
+                videofile = _downloadVideoFile(sourceflavor)
 
-              # Use this method to upload large files
-              #s3client.upload_file(videofile, s3bucketname, entry.id)
-              #  Apparently you can get the client from the resource this way
-              s3resource.meta.client.upload_file(videofile, s3bucketname, entry.id)
+              if not dryrun:
+                s3resource.meta.client.upload_file(videofile, s3bucketname, entry.id)
 
 # Catch/handle exceptions???
 # Integrity check???
 
               _addTag(entry, archivedtag)
 
-# Delete local file
-              os.remove(downloaddir + "/tempvideofile")
+              # Delete local file
+              if not dryrun:
+                os.remove(downloaddir + "/tempvideofile")
 
-# Delete source flavor
-              client.flavorAsset.delete(sourceflavor.id)
+              # Delete all flavors including source
+              _deleteEntryFlavors(entry.id, True)
 
               uploadPlaceholder(entry.id)
 
@@ -232,18 +264,19 @@ def archiveFlavors():
 
 
 def _downloadVideoFile(sourceflavor):
-          #print("\n".join(map(str, flavorassets)))
+    #print("\n".join(map(str, flavorassets)))
 
-          # Get the Download URL of the source video
-          src_url = client.flavorAsset.getUrl(sourceflavor.id)
-          logging.info("ID of src = %s" % sourceflavor.id)
-          logging.info("URL of src = %s" % src_url)
+    # Get the Download URL of the source video
+    src_url = client.flavorAsset.getUrl(sourceflavor.id)
+    logging.debug("Downloading source flavor = %s from %s" % (sourceflavor.id, src_url))
 
-          # Download the source video
-          filepath = downloaddir + "/tempvideofile"
-          urllib.urlretrieve (src_url, filepath)
+    # Download the source video
+    filepath = downloaddir + "/tempvideofile"
 
-          return filepath
+    if not dryrun:
+        urllib.urlretrieve (src_url, filepath)
+
+    return filepath
 
 
 def _S3ObjectExists(s3, bucketname, filename):
@@ -271,11 +304,14 @@ def restoreVideo(entryid):
   s3resource = boto3.resource('s3')
 
   logging.debug("Downloading video from S3")
-  s3resource.meta.client.download_file(s3bucketname, entryid, filepath)
+
+  if not dryrun:
+    s3resource.meta.client.download_file(s3bucketname, entryid, filepath)
 
   _uploadVideo(filepath)
 
-  os.remove(filepath)
+  if not dryrun:
+    os.remove(filepath)
 
 # Delete file from S3??
 
@@ -284,7 +320,7 @@ def restoreVideo(entryid):
 
 def uploadPlaceholder(entryid):
 
-    logging.debug("Uploading placeholder video for entry: " % (entryid))
+    logging.debug("Uploading placeholder video for entry: %s" % (entryid))
     _uploadVideo(entryid, placeholder_file_path)
 
 
@@ -292,18 +328,20 @@ def _uploadVideo(entryid, filepath):
 
   # Upload the file to Kaltura and re-link it to the media entry
 
-  logging.debug("Uploading video to Kaltura")
-  uploadToken = KalturaUploadToken();
-  uploadToken = client.uploadToken.add(uploadToken);
+  logging.debug("Uploading video to Kaltura entry: %s" % (entryid))
 
-  ulfile = file(filepath);
+  if not dryrun:
+      uploadToken = KalturaUploadToken()
+      uploadToken = client.uploadToken.add(uploadToken)
 
-  client.uploadToken.upload(uploadToken.id, ulfile);
+      ulfile = file(filepath)
 
-  uploadedFileTokenResource = KalturaUploadedFileTokenResource();
-  uploadedFileTokenResource.token = uploadToken.id;
+      client.uploadToken.upload(uploadToken.id, ulfile)
 
-  client.media.addContent(entryid, uploadedFileTokenResource);
+      uploadedFileTokenResource = KalturaUploadedFileTokenResource()
+      uploadedFileTokenResource.token = uploadToken.id
+
+      client.media.addContent(entryid, uploadedFileTokenResource)
 
 
 #######
@@ -312,22 +350,35 @@ def _uploadVideo(entryid, filepath):
 
 if __name__ == '__main__':
 
-  # Check configuration
-  checkConfig()
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--dryrun", help="Do not make any changes",
+                        action="store_true")
+    args = parser.parse_args()
 
-  ks = client.session.start(secret, userId, ktype, partnerId, expiry, privileges)
-  client.setKs(ks)
+    if args.dryrun:
+        dryrun = True
 
-  #deleteFlavors()
+    # Check configuration
+    checkConfig()
 
-  #archiveFlavors()
+    try:
+        ks = client.session.start(secret, userId, ktype, partnerId, expiry, privileges)
+        client.setKs(ks)
+    except KalturaException as e:
+        logging.fatal("KalturaException message: %s" % (e.message))
+        logging.fatal("Exiting immediately")
+        exit(errno.EACCES)
 
-  # Initiate retrieval from Glacier before being able to restore
-  # See https://thomassileo.name/blog/2012/10/24/getting-started-with-boto-and-glacier/
+    deleteFlavors()
 
-  uploadPlaceholder("1_6cwwzio0")
+    archiveFlavors()
 
-  #restoreVideo("1_6cwwzio0")
+    # Initiate retrieval from Glacier before being able to restore
+    # See https://thomassileo.name/blog/2012/10/24/getting-started-with-boto-and-glacier/
 
-  client.session.end()
+    #uploadPlaceholder("1_6cwwzio0", placeholder_file_path)
+
+    #restoreVideo("1_6cwwzio0")
+
+    client.session.end()
